moved createPollingConfigManager to utils.
